Add the possibility to set the working directory for the launched process.
The child process is now using chdir before calling execve. Since chdir may fail, a new pipe is created between the parent and the child to report the errno value to the notifier. This new pipe is then also used in case of dup2 errors.

Two unit tests are added to test a successful and a failing chdir. The _ProcessClient class in the tests has been generalized to be able to expect any kind of process error, and not just handle the special case for _TestFileExec.